:white_check_mark: Chore(Customer) Adding integration test for post
Implementing an integration test using testcontainers and rest-assured to verify that the creation of a new customer works as expected. The testcontainer setup is kept 
as simple as possible, with the static initialization in a companion object, as this is just a simple test case for playing around with testcontainers. In a real world project, one could abstract this away, into a central testcontainer setup class.